memenator/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coverage==6.5.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
factory_boy==3.2.0
pytest-django==4.5.2
setuptools>=65.0.0
drf-spectacular>=0.27.2
Pillow==10.0.0
imgurpython==1.1.7





